+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+#### Fixed
+- Default imports are removed regardless if they are used or not ([#149](https://github.com/buehler/typescript-hero/issues/149))
 
 ## [0.13.2]
 #### Fixed